refactor: user 도메인 clean Architecture + layered 방식으로 변경
