+++ src/user/domain/interface/use-case/user.use-case.ts
@@ -1,0 +1,7 @@
+import { UserDomain } from '../../model/user.domain';
+
+export const signUpUseCaseSymbol = Symbol.for('SignUpUseCase');
+
+export interface SignUpUseCase {
+  execute(name: string): Promise<UserDomain>;
+}
