+++ src/user/infrastructure/repository/user.repository.impl.ts
@@ -1,7 +1,7 @@
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserEntity } from '../entity/user.entity';
 import { Repository } from 'typeorm';
-import { UserRepository } from 'src/user/domain/interface/user.repository';
+import { UserRepository } from 'src/user/domain/interface/repository/user.repository';
 import { UserDomain } from 'src/user/domain/model/user.domain';
 import { Injectable } from '@nestjs/common';
 import { UserMapper } from 'src/user/domain/mapper/user.mapper';
