+++ src/user/presentation/controller/user.controller.ts
@@ -1,14 +1,14 @@
 import { Body, Controller, Inject, Post } from '@nestjs/common';
 import { SignUpUserRequest } from 'src/user/application/dto/request/sign-up-user.request';
 import { UserResponse } from 'src/user/application/dto/response/user.response';
-import { UserService, userServiceSymbol } from 'src/user/domain/interface/user.service';
+import { SignUpUseCase, signUpUseCaseSymbol } from 'src/user/domain/interface/use-case/user.use-case';
 
 @Controller('/users')
 export class UserController {
-  constructor(@Inject(userServiceSymbol) private readonly userService: UserService) {}
+  constructor(@Inject(signUpUseCaseSymbol) private readonly signUpUseCase: SignUpUseCase) {}
 
   @Post()
   async signUp(@Body() request: SignUpUserRequest): Promise<UserResponse> {
-    return this.userService.signUp(request.name);
+    return this.signUpUseCase.execute(request.name);
   }
 }
